third_party/WebKit/LayoutTests/http/tests/security/mixedContent/insecure-fetch-on-dedicated-worker-blocked.https.html

<!DOCTYPE html>
<p>This test starts a dedicated worker that loads insecure data via Fetch API.
We should trigger a mixed content callback and block the load.</p>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
if (window.testRunner) {
  testRunner.overridePreference("WebKitAllowRunningInsecureContent", false);
}

async_test(t => {
  var worker = new Worker("./resources/insecure-fetch-in-dedicated-worker.js");
  worker.addEventListener("message", t.step_func_done(e => {
    assert_equals(e.data, "LOAD FAILED");
  }));
}, "Insecure resource loads inside Worker failes when" +
   " 'AllowRunningInsecureContent' is not set.");

</script>
